requirements.txt
# Current dependencies
torch==2.5.1
transformers==4.36.2
sentence-transformers==2.2.2
chromadb==0.4.22
huggingface-hub==0.19.4
loguru==0.7.2
pyyaml==6.0.1

# Add free LLM dependencies
accelerate==0.25.0
bitsandbytes==0.41.1